backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pdfplumber>=0.11.4
PyPDF2>=3.0.0
python-docx>=1.1.2
spacy>=3.7.0
pydantic>=2.9.0
pydantic[email]>=2.9.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
transformers>=4.41.0
torch>=1.11.0
requests>=2.31.0
openai>=1.0.0
pymongo>=4.6.0
bcrypt>=4.1.0
pyjwt>=2.8.0
